+++ re2c/src/util/substr.h
@@ -1,7 +1,6 @@
 #ifndef _RE2C_UTIL_SUBSTR_
 #define _RE2C_UTIL_SUBSTR_
 
-#include "src/util/c99_stdint.h"
 #include "src/util/forbid_copy.h"
 
 namespace re2c
@@ -12,14 +11,14 @@ class SubStr
 public:
 	const char * str;
 	const char * const org;
-	uint32_t len;
+	size_t len;
 
-	inline SubStr (const char * s, uint32_t l)
+	inline SubStr (const char * s, size_t l)
 		: str (s)
 		, org (s)
 		, len (l)
 	{}
-	inline uint32_t ofs () const
+	inline ptrdiff_t ofs () const
 	{
 		return str - org;
 	}
